Moved project classes to "libs" directory
